Append suffix to SIMD implementation specific files
