+++ OgreMain/include/Math/Array/C/OgreArrayQuaternionC.inl
