+++ OgreMain/include/Math/Array/C/OgreArraySphereC.inl
