+++ OgreMain/include/Math/Array/C/OgreBooleanMaskC.inl
