+++ OgreMain/include/Math/Array/NEON/Single/OgreArrayMatrix4NEON.h
@@ -253,6 +253,6 @@ namespace Ogre
 
 }
 
-#include "OgreArrayMatrix4.inl"
+#include "OgreArrayMatrix4NEON.inl"
 
 #endif
