+++ OgreMain/include/Math/Array/NEON/Single/OgreArrayQuaternionNEON.inl
