+++ OgreMain/include/Math/Array/NEON/Single/OgreArraySphereNEON.inl
