+++ OgreMain/include/Math/Array/NEON/Single/OgreBooleanMaskNEON.inl
