+++ OgreMain/include/Math/Array/OgreArraySphere.h
@@ -33,18 +33,18 @@ THE SOFTWARE.
 
 #if __OGRE_HAVE_SSE
     #if OGRE_DOUBLE_PRECISION == 1
-        #include "SSE2/Double/OgreArraySphere.h"
+        #include "SSE2/Double/OgreArraySphereSSE2.h"
     #else
-        #include "SSE2/Single/OgreArraySphere.h"
+        #include "SSE2/Single/OgreArraySphereSSE2.h"
     #endif
 #elif __OGRE_HAVE_NEON
     #if OGRE_DOUBLE_PRECISION == 1
         #error Double precision with SIMD on ARM is not supported
     #else
-        #include "NEON/Single/OgreArraySphere.h"
+        #include "NEON/Single/OgreArraySphereNEON.h"
     #endif
 #else
-    #include "C/OgreArraySphere.h"
+    #include "C/OgreArraySphereC.h"
 #endif
 
 #endif
