+++ OgreMain/include/Math/Array/OgreArrayVector3.h
@@ -33,18 +33,18 @@ THE SOFTWARE.
 
 #if __OGRE_HAVE_SSE
     #if OGRE_DOUBLE_PRECISION == 1
-        #include "SSE2/Double/OgreArrayVector3.h"
+        #include "SSE2/Double/OgreArrayVector3SSE2.h"
     #else
-        #include "SSE2/Single/OgreArrayVector3.h"
+        #include "SSE2/Single/OgreArrayVector3SSE2.h"
     #endif
 #elif __OGRE_HAVE_NEON
     #if OGRE_DOUBLE_PRECISION == 1
         #error Double precision with SIMD on ARM is not supported
     #else
-        #include "NEON/Single/OgreArrayVector3.h"
+        #include "NEON/Single/OgreArrayVector3NEON.h"
     #endif
 #else
-    #include "C/OgreArrayVector3.h"
+    #include "C/OgreArrayVector3C.h"
 #endif
 
 #endif
