+++ OgreMain/include/Math/Array/OgreBooleanMask.h
@@ -33,18 +33,18 @@ THE SOFTWARE.
 
 #if __OGRE_HAVE_SSE
     #if OGRE_DOUBLE_PRECISION == 1
-        #include "SSE2/Double/OgreBooleanMask.h"
+        #include "SSE2/Double/OgreBooleanMaskSSE2.h"
     #else
-        #include "SSE2/Single/OgreBooleanMask.h"
+        #include "SSE2/Single/OgreBooleanMaskSSE2.h"
     #endif
 #elif __OGRE_HAVE_NEON
     #if OGRE_DOUBLE_PRECISION == 1
         #error Double precision with SIMD on ARM is not supported
     #else
-        #include "NEON/Single/OgreBooleanMask.h"
+        #include "NEON/Single/OgreBooleanMaskNEON.h"
     #endif
 #else
-    #include "C/OgreBooleanMask.h"
+    #include "C/OgreBooleanMaskC.h"
 #endif
 
 #endif
