+++ OgreMain/include/Math/Array/SSE2/Single/OgreArrayMatrix4SSE2.h
@@ -238,6 +238,6 @@ namespace Ogre
 
 }
 
-#include "OgreArrayMatrix4.inl"
+#include "OgreArrayMatrix4SSE2.inl"
 
 #endif
