+++ OgreMain/include/Math/Array/SSE2/Single/OgreArrayQuaternionSSE2.h
@@ -37,8 +37,6 @@ THE SOFTWARE.
 #include "Math/Array/OgreMathlib.h"
 #include "Math/Array/OgreArrayVector3.h"
 
-#include "OgreArrayQuaternion.h"
-
 namespace Ogre
 {
     /** \addtogroup Core
@@ -253,6 +251,6 @@ namespace Ogre
 
 }
 
-#include "OgreArrayQuaternion.inl"
+#include "OgreArrayQuaternionSSE2.inl"
 
 #endif
