+++ OgreMain/include/Math/Array/SSE2/Single/OgreArrayQuaternionSSE2.inl
