+++ OgreMain/include/Math/Array/SSE2/Single/OgreArraySphereSSE2.inl
