+++ OgreMain/include/Math/Array/SSE2/Single/OgreArrayVector3SSE2.h
@@ -373,6 +373,6 @@ namespace Ogre
 
 }
 
-#include "OgreArrayVector3.inl"
+#include "OgreArrayVector3SSE2.inl"
 
 #endif
