+++ OgreMain/include/Math/Simple/C/OgreAabbC.h
@@ -173,6 +173,6 @@ namespace Ogre
 
 }
 
-#include "OgreAabb.inl"
+#include "OgreAabbC.inl"
 
 #endif
